+++ lib/property_value.dart
@@ -8,19 +8,19 @@ part of './mvvm.dart';
 ///
 class ValueViewModelProperty<TValue> extends BindableProperty<TValue> {
   ValueViewModelProperty._(
-      Object propertyKey, ValueNotifier<TValue> valueNotifier,
+      Object propertyKey, BindableValueNotifier<TValue> valueNotifier,
       {PropertyValueChanged<TValue> valueChanged})
       : super(propertyKey, valueNotifier, valueChanged: valueChanged);
 
   /// ValueViewModelProperty
   ValueViewModelProperty(Object propertyKey,
       {PropertyValueChanged<TValue> valueChanged, TValue value})
-      : this._(propertyKey, ValueNotifier<TValue>(value),
+      : this._(propertyKey, BindableValueNotifier<TValue>(value),
             valueChanged: valueChanged);
 
   /// ValueViewModelProperty.from
   ValueViewModelProperty.from(
-      Object propertyKey, ValueNotifier<TValue> valueNotifier,
+      Object propertyKey, BindableValueNotifier<TValue> valueNotifier,
       {PropertyValueChanged<TValue> valueChanged})
       : this._(propertyKey, valueNotifier, valueChanged: valueChanged);
 }
@@ -30,7 +30,9 @@ mixin ValueViewModelMixin on _ViewModelBase {
   /// 创建一个值属性
   ///
   /// [propertyKey] 指定属性键
+  /// 
   /// [valueChanged] 指定属性值变更后的回调方法
+  /// 
   /// [initial] 指定初始值
   ///
   BindableProperty<TValue> propertyValue<TValue>(Object propertyKey,